backend/requirements.txt
fastapi==0.119.0
uvicorn[standard]==0.38.0
httpx==0.28.1
pytest==8.4.2
PyJWT==2.10.1
